requirements.txt
langchain==0.0.339
gradio==4.5.0
python-dotenv==1.0.0
openai==1.3.4